pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mechanicsdsl-core"
version = "2.0.0"
description = "A Domain-Specific Language and Transpiler for Classical Mechanics"
authors = [
  { name="Noah Parsons", email="nomapa223@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
keywords = [
    "physics", "simulation", "mechanics", "dsl", "lagrangian",
    "hamiltonian", "code-generation", "scientific-computing",
    "embedded", "raspberry-pi", "arm", "gpu", "cuda"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: C++",
    "Programming Language :: Rust",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Compilers",
    "Topic :: System :: Emulators",
    "Topic :: System :: Hardware",
]
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "sympy>=1.8",
    "matplotlib>=3.4.0",
]

[project.optional-dependencies]
test = ["pytest", "pytest-cov", "hypothesis>=6.0"]
codegen = ["pybind11"]
jit = ["numba>=0.56.0"]
typing = ["mypy", "types-setuptools"]
# New v1.6.0 optional dependencies
jax = ["jax>=0.4.0", "jaxlib>=0.4.0", "diffrax>=0.4.0"]
jax-gpu = ["jax[cuda12]>=0.4.0", "diffrax>=0.4.0"]
server = ["fastapi>=0.100.0", "uvicorn>=0.22.0", "websockets>=11.0"]
jupyter = ["ipywidgets>=8.0.0", "ipython>=8.0.0"]
lsp = ["pygls>=1.0.0", "lsprotocol>=2023.0.0"]
openmao = ["openmdao>=3.25.0"]
embedded = ["RPi.GPIO>=0.7.0; platform_machine == 'aarch64' or platform_machine == 'armv7l'"]
arm = ["numpy>=1.20.0"]  # ARM-optimized builds
dev = [
    "pytest",
    "pytest-cov",
    "hypothesis>=6.0",
    "mypy",
    "types-setuptools",
    "black",
    "flake8",
    "isort",
    "bandit",
]
all = [
    "numba>=0.56.0", 
    "pybind11", 
    "hypothesis>=6.0", 
    "mypy",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "ipywidgets>=8.0.0",
    "pygls>=1.0.0",
]

[project.urls]
"Repository" = "https://github.com/MechanicsDSL/mechanicsdsl"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu_required: marks tests that require GPU/CUDA (deselect with '-m \"not gpu_required\"')",
    "physics_validation: marks tests that validate physics correctness",
    "performance: marks performance/benchmark tests",
    "integration: marks integration tests",
    "numba_required: marks tests that require numba JIT compiler",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312", "py313"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true

